// client/da/uml-server/src/language-server/generated/ast.ts
export type Concurrency = 'CONCURRENT' | 'GUARDED' | 'SEQUENTIAL';

export function isConcurrency(item: unknown): item is Concurrency {
    return item === 'SEQUENTIAL' || item === 'GUARDED' || item === 'CONCURRENT';
}

export type DataTypeReference = Class | DataType | Enumeration | Interface | PrimitiveType;

export const DataTypeReference = 'DataTypeReference';

export function isDataTypeReference(item: unknown): item is DataTypeReference {
    return reflection.isInstance(item, DataTypeReference);
}

export type EffectType = 'CREATE' | 'DELETE' | 'READ' | 'UPDATE';
// ...
export type ParameterDirection = 'IN' | 'INOUT' | 'OUT' | 'RETURN';
// ...
export type Visibility = 'PACKAGE' | 'PRIVATE' | 'PROTECTED' | 'PUBLIC';

export function isVisibility(item: unknown): item is Visibility {
    return item === 'PUBLIC' || item === 'PRIVATE' || item === 'PROTECTED' || item === 'PACKAGE';
}

export interface Class extends AstNode {
    readonly $type: 'AbstractClass' | 'Class';
    __id: string
    isAbstract: boolean
    isActive: boolean
    name: string
    operations: Array<Operation>
    properties: Array<Property>
    skip: boolean
    visibility: Visibility
}

export const Class = 'Class';

export function isClass(item: unknown): item is Class {
    return reflection.isInstance(item, Class);
}

export interface ClassDiagram extends AstNode {
    readonly $container: Diagram;
    readonly $type: 'ClassDiagram';
    __id: string
    diagramType: 'CLASS'
    entities: Array<Entity>
    relations: Array<Relation>
}

export const ClassDiagram = 'ClassDiagram';

export function isClassDiagram(item: unknown): item is ClassDiagram {
    return reflection.isInstance(item, ClassDiagram);
}

export interface DataType extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'DataType';
    __id: string
    isAbstract: boolean
    name: string
    operations: Array<Operation>
    properties: Array<Property>
    visibility?: Visibility
}

export const DataType = 'DataType';
// ...
export interface Interface extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'Interface';
    __id: string
    name: string
    operations: Array<Operation>
    properties: Array<Property>
}

export const Interface = 'Interface';
// ...
export interface Operation extends AstNode {
    readonly $container: AbstractClass | Class | DataType | Interface;
    readonly $type: 'Operation';
    __id: string
    concurrency?: Concurrency
    isAbstract: boolean
    isQuery: boolean
    isStatic: boolean
    name: string
    parameters: Array<Parameter>
    visibility?: Visibility
}

export const Operation = 'Operation';

export function isOperation(item: unknown): item is Operation {
    return reflection.isInstance(item, Operation);
}

export interface Package extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'Package';
    __id: string
    entities: Array<Entity>
    name: string
    uri?: string
    visibility?: Visibility
}

export const Package = 'Package';

export function isPackage(item: unknown): item is Package {
    return reflection.isInstance(item, Package);
}

export interface PackageDiagram extends AstNode {
    readonly $container: Diagram;
    readonly $type: 'PackageDiagram';
    __id: string
    diagramType: 'PACKAGE'
    entities: Array<Entity>
    relations: Array<Relation>
}

export const PackageDiagram = 'PackageDiagram';

export function isPackageDiagram(item: unknown): item is PackageDiagram {
    return reflection.isInstance(item, PackageDiagram);
}

export interface Parameter extends AstNode {
    readonly $container: Operation;
    readonly $type: 'Parameter';
    __id: string
    direction?: ParameterDirection
    effect?: EffectType
    isException: boolean
    isOrdered: boolean
    isStream: boolean
    isUnique: boolean
    multiplicity?: string
    name: string
    parameterType?: Reference<DataTypeReference>
    visibility?: Visibility
}

export const Parameter = 'Parameter';
// ...
export interface Property extends AstNode {
    readonly $container: AbstractClass | Class | DataType | Interface;
    readonly $type: 'Property';
    __id: string
    aggregation?: AggregationType
    isDerived: boolean
    isDerivedUnion: boolean
    isOrdered: boolean
    isReadOnly: boolean
    isStatic: boolean
    isUnique: boolean
    multiplicity?: string
    name: string
    propertyType?: Reference<DataTypeReference>
    visibility?: Visibility
}

export const Property = 'Property';
// ...
export interface AbstractClass extends Class {
    readonly $type: 'AbstractClass';
    __id: string
    isAbstract: boolean
    isActive: boolean
    label: string
    name: string
    operations: Array<Operation>
    properties: Array<Property>
    skip: boolean
    visibility: Visibility
}

export const AbstractClass = 'AbstractClass';